requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
exceptiongroup==1.3.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.22.0
idna==3.10
numpy==2.0.2
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules
pycparser
pydantic
pydantic_core
PyMuPDF
pyparsing
pypdfium2
python-dotenv
redis
requests
rsa
sniffio
starlette
tqdm
typing-inspection
typing_extensions
uritemplate
urllib3
uvicorn
uvicorn[standard]
python-multipart

aiohttp
# Utilities
python-jose
passlib
python-slugify
shortuuid
# Development
pytest
pytest-asyncio
black
isort
flake8